+++ src/CfdiDataBuilder.php
@@ -17,11 +17,9 @@ class CfdiDataBuilder
 {
     use CastToStringTrait;
 
-    /** @var XmlResolver */
-    private $xmlResolver;
+    private XmlResolver $xmlResolver;
 
-    /** @var XsltBuilderInterface */
-    private $xsltBuilder;
+    private XsltBuilderInterface $xsltBuilder;
 
     public function __construct()
     {
@@ -51,10 +49,6 @@ class CfdiDataBuilder
         return $this->xsltBuilder;
     }
 
-    /**
-     * @param NodeInterface<NodeInterface> $comprobante
-     * @return CfdiData
-     */
     public function build(NodeInterface $comprobante): CfdiData
     {
         return new CfdiData(
@@ -64,10 +58,6 @@ class CfdiDataBuilder
         );
     }
 
-    /**
-     * @param NodeInterface<NodeInterface> $comprobante
-     * @return string
-     */
     public function createTfdSourceString(NodeInterface $comprobante): string
     {
         $tfd = $comprobante->searchNode('cfdi:Complemento', 'tfd:TimbreFiscalDigital');
@@ -81,10 +71,6 @@ class CfdiDataBuilder
         );
     }
 
-    /**
-     * @param NodeInterface<NodeInterface> $comprobante
-     * @return string
-     */
     public function createQrUrl(NodeInterface $comprobante): string
     {
         $parameters = new RequestParameters(